+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/handler/ReverseProxyCallback.java
@@ -30,7 +30,6 @@ public class ReverseProxyCallback extends AbstractTigerRouteCallback {
         } else {
             port = targetUri.getPort();
         }
-
     }
 
     @Override